Develop

+++ src/services/ProductService.ts
@@ -1,8 +1,14 @@
-import { CreateProductDto } from "../dto/product/CreateProductDto";
-import UpdateProductDto from "../dto/product/UpdateProductDto";
+import ProductDto from "../dto/ProductDto";
 import HttpException from "../exceptions/HttpException";
 import Product from "../models/Product";
-import IProductService from "./interfaces/IProductService";
+
+interface IProductService {
+    createProduct(data: ProductDto): void;
+    updateProduct(id: String, data: ProductDto): Promise<void>;
+    deleteProduct(id: String): Promise<void>;
+    getProduct(id: String): Promise<Object>;
+    getActiveProducts(): void;
+}
 
 class ProductService implements IProductService {
 
@@ -15,19 +21,22 @@ class ProductService implements IProductService {
         return product;
     }
 
-    async getActiveProducts(): Promise<Object> {
+    async getActiveProducts() {
         const products = await Product.find({ status: true }).exec();
         return products;
     }
 
-    async createProduct(data: CreateProductDto): Promise<Object> {
+    async createProduct(data: ProductDto)  {
         const product = await Product.create(data);
         return product;
     }
 
-    async updateProduct(id: String, data: UpdateProductDto): Promise<void> {
+    async updateProduct(id: String, data: ProductDto): Promise<void> {
         
-        await this.checkProductExists(id);
+        const productExists = await this.checkProductExists(id);
+        if(!productExists) {
+            throw new Error('Product not found');
+        }
 
         const productUpdated = await Product.updateOne({ _id: id }, data);
         if(!productUpdated.modifiedCount) {
@@ -37,7 +46,10 @@ class ProductService implements IProductService {
 
     async deleteProduct(id: String): Promise<void> {
 
-        await this.checkProductExists(id);
+        const productExists = await this.checkProductExists(id);
+        if(!productExists) {
+            throw new Error('Product not found');
+        }
 
         const deletedProduct = await Product.deleteOne({ _id: id });
         if(deletedProduct.deletedCount === 0) {
@@ -45,12 +57,14 @@ class ProductService implements IProductService {
         }
     }
 
-    private async checkProductExists(id: String) {
+    private async checkProductExists(id: String): Promise<boolean> {
         const productExists = await Product.findById({ _id: id });
-        if(!productExists) {
-            throw new HttpException(404, 'Product not found');
+        if(productExists) {
+            return true;
         }
+
+        return false;
     }
 }
 
-export default new ProductService()
+export default ProductService;
